Website Session control
The website now knows who the logged user is and shows content based on that.

Improved User roles assignment. Now the entity User has a List of roles instead of a single String representing its role. This way, it is easier to nest permissions.

+++ src/main/java/es/codeurjc/holamundo/controller/ReviewSectionController.java
@@ -4,7 +4,9 @@ import es.codeurjc.holamundo.entity.Review;
 import es.codeurjc.holamundo.repository.BookRepository;
 import es.codeurjc.holamundo.repository.ReviewRepository;
 import es.codeurjc.holamundo.repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -12,7 +14,7 @@ import org.springframework.web.bind.annotation.*;
 @Controller
 public class ReviewSectionController {
     // Change this when we have session control
-    private String testingCurrentUsername = "FanBook_785"; //This is the username of the current user. This is just for testing purposes
+    private String testingCurrentUsername;
 
     @Autowired
     private ReviewRepository reviewRepository;
@@ -27,7 +29,13 @@ public class ReviewSectionController {
     public String addReview(Model model, @PathVariable int id,
                             @RequestParam("rating") int rating,
                             @RequestParam("reviewTitle") String title,
-                            @RequestParam("comment") String comment) {
+                            @RequestParam("comment") String comment, HttpServletRequest request) {
+
+        Authentication authentication = (Authentication) request.getUserPrincipal();
+        if (authentication != null) {
+            testingCurrentUsername = authentication.getName();
+        }
+
         System.out.println("Adding review" + title + " " + rating + " " + comment + " " + id);
         // Add a review to the database
         Review review = new Review(title, userRepository.findByUsername(testingCurrentUsername), rating, comment, bookRepository.findByID(id));
@@ -36,10 +44,20 @@ public class ReviewSectionController {
     }
 
     @DeleteMapping("/book/{id}/deleteReview/{reviewID}")
-    public String deleteReview(Model model, @PathVariable int id, @PathVariable long reviewID) {
-        System.out.println("Deleting review" + reviewID);
-        // Delete a review from the database
-        reviewRepository.deleteById(reviewID);
+    public String deleteReview(Model model, @PathVariable int id, @PathVariable long reviewID, HttpServletRequest request) {
+
+        // Check that the user is the owner of the review
+        Authentication authentication = (Authentication) request.getUserPrincipal();
+
+        // Check role (if admin, allow to delete any review)
+        if (authentication != null) {
+            testingCurrentUsername = authentication.getName();
+
+            if (reviewRepository.findByID(reviewID).getUser().getUsername().equals(testingCurrentUsername) || userRepository.findByUsername(testingCurrentUsername).getRole().contains("ADMIN")) {
+                reviewRepository.deleteById(reviewID);
+            }
+        }
+
         return "redirect:/book/" + id;
     }
